+++ cmd/cortextool/pkg/chunk/gcp/bigtable_delete.go
@@ -11,7 +11,7 @@ import (
 	ot "github.com/opentracing/opentracing-go"
 	"github.com/sirupsen/logrus"
 
-	chunkTool "github.com/grafana/cortextool/pkg/chunk"
+	chunkTool "github.com/grafana/cortex-tools/pkg/chunk"
 )
 
 const (