+++ cmd/cortextool/pkg/chunk/storage/factory.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/cortexproject/cortex/pkg/chunk/storage"
 
-	"github.com/grafana/cortextool/pkg/chunk"
-	"github.com/grafana/cortextool/pkg/chunk/gcp"
+	"github.com/grafana/cortex-tools/pkg/chunk"
+	"github.com/grafana/cortex-tools/pkg/chunk/gcp"
 )
 
 // NewChunkScanner makes a new table client based on the configuration.